fatura-app/requirements.txt
Werkzeug==2.3.7
Flask==2.3.3
mysql-connector-python==8.1.0
pymongo